Description: 	Switch 'unsigned' variable to 'size_t' for H5Pget_filter_by_id2() call
Tested on:
        FreeBSD/32 6.3 (duty) in debug mode
        FreeBSD/64 6.3 (liberty) w/C++ & FORTRAN, in debug mode
        Linux/32 2.6 (kagiso) w/PGI compilers, w/C++ & FORTRAN, w/threadsafe,
                                in debug mode
        Linux/64-amd64 2.6 (smirom) w/Intel compilers w/default API=1.6.x,
                                w/C++ & FORTRAN, in production mode
        Solaris/32 2.10 (linew) w/deprecated symbols disabled, w/C++ & FORTRAN,
                                w/szip filter, in production mode
        Linux/64-ia64 2.6 (cobalt) w/Intel compilers, w/C++ & FORTRAN,
                                in production mode
        Linux/64-ia64 2.4 (tg-login3) w/parallel, w/FORTRAN, in production mode
        Linux/64-amd64 2.6 (abe) w/parallel, w/FORTRAN, in production mode
        Mac OS X/32 10.5.8 (amazon) in debug mode
        Mac OS X/32 10.5.8 (amazon) w/C++ & FORTRAN, w/threadsafe,
                                in production mode

+++ branches/compress_group_fheap/test/links.c
@@ -8602,7 +8602,7 @@ static int
 link_filters(hid_t fapl, hbool_t new_format)
 {
     hid_t       fid = -1, gid1 = -1, gid2 = -1, gcpl1 = -1, gcpl2 = -1, fcpl = -1;
-    unsigned    cd_nelmts = 1;
+    size_t      cd_nelmts = 1;
     unsigned    cd_value = 2112;
     unsigned    cd_value_out;
     unsigned    flags_out;
